cast rdev to uint64

+++ process_posix.go
@@ -40,7 +40,7 @@ func getTerminalMap() (map[uint64]string, error) {
 	for _, name := range termfiles {
 		stat := syscall.Stat_t{}
 		syscall.Stat(name, &stat)
-		rdev := stat.Rdev
+		rdev := uint64(stat.Rdev)
 		ret[rdev] = strings.Replace(name, "/dev", "", -1)
 	}
 	return ret, nil
